Embed tarball of QEMU's source inside container
Download and save the tarball of QEMU's source code on container
creation so we can use the container offline with no required downloads
unless a different version of QEMU is specified via `QEMU_VER` variable.

+++ entrypoint.sh
@@ -208,8 +208,12 @@ makeAppImage() {
 
 # Download QEMU's source code
 if [ ! -x ./configure ]; then
-	wget "https://download.qemu.org/qemu-${VERSION}.tar.xz" -O - \
-		| tar -xJv --strip-components=1 || exit
+	if [ -f "/qemu-${VERSION}.tar.xz" ]; then
+		tar -xJv --strip-components=1 -f "/qemu-${VERSION}.tar.xz" || exit
+	else
+		wget "https://download.qemu.org/qemu-${VERSION}.tar.xz" -O - \
+			| tar -xJv --strip-components=1 || exit
+	fi
 elif [ -f ./VERSION ]; then
 	VERSION=$(cat ./VERSION)
 fi
